rename to OStream
