+++ exporters/ostream/BUILD
@@ -1,12 +1,12 @@
 package(default_visibility = ["//visibility:public"])
 
 cc_library(
-    name = "stream_metrics_exporter",
+    name = "ostream_metrics_exporter",
     srcs = [
         "src/metrics_exporter.cc",
     ],
     hdrs = [
-        "include/opentelemetry/exporters/stream/metrics_exporter.h",
+        "include/opentelemetry/exporters/ostream/metrics_exporter.h",
     ],
     strip_include_prefix = "include",
     deps = [
@@ -15,10 +15,10 @@ cc_library(
  )
 
 cc_test(
-    name = "stream_metrics_test",
-    srcs = ["test/stream_metrics_test.cc"],
+    name = "ostream_metrics_test",
+    srcs = ["test/ostream_metrics_test.cc"],
     deps = [
-        ":stream_metrics_exporter",
+        ":ostream_metrics_exporter",
         "@com_google_googletest//:gtest_main",
     ],
  )
